+++ tests/test_strspn_s.c
@@ -183,7 +183,7 @@ int test_strspn_s (void)
         errs++;
     }
     std_count = strspn(str1, str2);
-    if (count != std_count) {
+    if ((int)count != std_count) {
         debug_printf("%s %u  count=%d  std_count=%d  rc=%d \n",
                      __FUNCTION__, __LINE__,  (int)count, std_count, rc);
         errs++;
@@ -200,7 +200,7 @@ int test_strspn_s (void)
         errs++;
     }
     std_count = strspn(str1, str2);
-    if (count != std_count) {
+    if ((int)count != std_count) {
         debug_printf("%s %u  count=%d  std_count=%d  rc=%d \n",
                      __FUNCTION__, __LINE__,  (int)count, std_count, rc);
         errs++;
@@ -218,7 +218,7 @@ int test_strspn_s (void)
         errs++;
     }
     std_count = strspn(str1, str2);
-    if (count != std_count) {
+    if ((int)count != std_count) {
         debug_printf("%s %u  count=%d  std_count=%d  rc=%d \n",
                      __FUNCTION__, __LINE__,  (int)count, std_count, rc);
         errs++;
@@ -235,7 +235,7 @@ int test_strspn_s (void)
         errs++;
     }
     std_count = strspn(str1, str2);
-    if (count != std_count) {
+    if ((int)count != std_count) {
         debug_printf("%s %u  count=%d  std_count=%d  rc=%d \n",
                      __FUNCTION__, __LINE__,  (int)count, std_count, rc);
         errs++;
@@ -252,7 +252,7 @@ int test_strspn_s (void)
         errs++;
     }
     std_count = strspn(str1, str2);
-    if (count != std_count) {
+    if ((int)count != std_count) {
         debug_printf("%s %u  count=%d  std_count=%d  rc=%d \n",
                      __FUNCTION__, __LINE__,  (int)count, std_count, rc);
         errs++;
@@ -269,7 +269,7 @@ int test_strspn_s (void)
         errs++;
     }
     std_count = strspn(str1, str2);
-    if (count != std_count) {
+    if ((int)count != std_count) {
         debug_printf("%s %u  count=%d  std_count=%d  rc=%d \n",
                      __FUNCTION__, __LINE__,  (int)count, std_count, rc);
         errs++;
